// frontend/pharmacy-app/src/enums/tableColumnKeys.ts
export enum ProcessedOrdersTableColumnKeys {
  ID = "id",
  CREATED_AT = "createdAt",
  PATIENT_MRN = "patientMrn",
  DRUG_NAME = "drugName",
  STATUS = "status",
  RE_RUN_ONCO_EMR = "reRunOncoEmr",
  RE_RUN_CMM = "reRunCmm",
  VIEW_CMM_INPUT = "viewCmmInput",
  REPORT_PRESCRIPTION_INACCURACY = "reportPrescriptionInaccuracy",
}